+++ adsdk/src/main/java/com/inner/adsdk/adloader/mopub/MopubLoader.java
@@ -499,7 +499,7 @@ public class MopubLoader extends AbstractSdkLoader {
                 if (getAdListener() != null) {
                     AdReward adReward = new AdReward();
                     adReward.setType(Constant.ECPM);
-                    int ecpm = 0;
+                    double ecpm = 0;
                     if (mPidConfig != null) {
                         ecpm = mPidConfig.getEcpm();
                     }
